bugfixing/renaming for libpolypartition
