+++ cob_3d_features/package.xml
@@ -15,6 +15,7 @@
 
   <buildtool_depend>catkin</buildtool_depend>
 
+  <build_depend>cmake_modules</build_depend>
   <build_depend>libopencv-dev</build_depend>
   <build_depend>boost</build_depend>
   <build_depend>cob_3d_mapping_common</build_depend>
@@ -22,6 +23,7 @@
   <build_depend>pcl_ros</build_depend>
   <build_depend>libvtk</build_depend>
   <build_depend>cob_3d_visualization</build_depend>
+  <build_depend>libpolypartition</build_depend>
 
   <run_depend>libopencv-dev</run_depend>
   <run_depend>boost</run_depend>
@@ -29,7 +31,9 @@
   <run_depend>libpcl-all-dev</run_depend>
   <run_depend>pcl_ros</run_depend>
   <run_depend>libvtk</run_depend>
-
+  <run_depend>cob_3d_visualization</run_depend>
+  <run_depend>libpolypartition</run_depend>
+  
 </package>
