+++ cob_3d_mapping_common/package.xml
@@ -19,7 +19,7 @@
   <build_depend>eigen_conversions</build_depend>
   <build_depend>libgpc</build_depend>
   <build_depend>pcl_conversions</build_depend>
-  <build_depend>polypartition</build_depend>
+  <build_depend>libpolypartition</build_depend>
   <build_depend>roscpp</build_depend>
 
   <!--run_depend>boost</run_depend-->
@@ -27,7 +27,7 @@
   <run_depend>eigen_conversions</run_depend>
   <run_depend>libgpc</run_depend>
   <run_depend>pcl_conversions</run_depend>
-  <run_depend>polypartition</run_depend>
+  <run_depend>libpolypartition</run_depend>
   <run_depend>roscpp</run_depend>
 </package>
