+++ cob_3d_segmentation/CMakeLists.txt
@@ -1,10 +1,7 @@
 cmake_minimum_required(VERSION 2.8.3)
 project(cob_3d_segmentation)
 
-## Find catkin macros and libraries
-## if COMPONENTS list like find_package(catkin REQUIRED COMPONENTS xyz)
-## is used, also find other catkin packages
-find_package(catkin REQUIRED COMPONENTS
+set(catkin_RUN_PACKAGES 
   actionlib
   actionlib_msgs
   cob_3d_features
@@ -13,19 +10,31 @@ find_package(catkin REQUIRED COMPONENTS
   cob_3d_mapping_msgs
   dynamic_reconfigure
   geometry_msgs
-  message_generation
   nodelet
   pcl_ros
-  polypartition
+  libpolypartition
   roscpp
   rosbag
+)
+
+set(catkin_BUILD_PACKAGES 
+	${catkin_RUN_PACKAGES}
+	cmake_modules
+	message_generation
+)
+
+## Find catkin macros and libraries
+## if COMPONENTS list like find_package(catkin REQUIRED COMPONENTS xyz)
+## is used, also find other catkin packages
+find_package(catkin REQUIRED COMPONENTS
+  ${catkin_BUILD_PACKAGES}
 )
 
 #find_package(TinyXML REQUIRED)
 find_package(VTK REQUIRED COMPONENTS vtkCommon vtkFiltering vtkRendering)
 find_package(Boost REQUIRED COMPONENTS program_options)
 find_package(OpenCV REQUIRED)
-find_package(PCL REQUIRED)
+#find_package(PCL REQUIRED)
 include(cmake/OptimizeForArchitecture.cmake)
 OptimizeForArchitecture()
 
@@ -69,24 +78,14 @@ catkin_package(
     cob_3d_segmentation
     cob_3d_segmentation_nodelets
   CATKIN_DEPENDS
-    actionlib
-    actionlib_msgs
-    cob_3d_features
-    cob_3d_mapping_common
-    cob_3d_mapping_filters
-    cob_3d_mapping_msgs
-    dynamic_reconfigure
-    geometry_msgs
-    nodelet
-    pcl_ros
-    roscpp
-    rosbag
+    ${catkin_RUN_PACKAGES}
+    message_runtime
   DEPENDS
 #    TinyXML
     VTK
     Boost
     OpenCV
-    PCL
+#    PCL
 )
