+++ cob_3d_segmentation/package.xml
@@ -15,6 +15,7 @@
 
   <buildtool_depend>catkin</buildtool_depend>
 
+  <build_depend>cmake_modules</build_depend>
   <build_depend>actionlib</build_depend>
   <build_depend>actionlib_msgs</build_depend>
   <build_depend>boost</build_depend>
@@ -24,14 +25,14 @@
   <build_depend>cob_3d_mapping_msgs</build_depend>
   <build_depend>dynamic_reconfigure</build_depend>
   <build_depend>geometry_msgs</build_depend>
-  <build_depend>libpcl-all-dev</build_depend>
+  <build_depend>libpcl-all</build_depend>
   <build_depend>libvtk</build_depend>
   <build_depend>message_generation</build_depend>
   <build_depend>nodelet</build_depend>
-  <build_depend>opencv2</build_depend>
+  <build_depend>libopencv-dev</build_depend>
   <!--build_depend>pcl</build_depend-->
   <build_depend>pcl_ros</build_depend>
-  <build_depend>polypartition</build_depend>
+  <build_depend>libpolypartition</build_depend>
   <build_depend>roscpp</build_depend>
   <build_depend>rosbag</build_depend>
 
@@ -44,14 +45,14 @@
   <run_depend>cob_3d_mapping_msgs</run_depend>
   <run_depend>dynamic_reconfigure</run_depend>
   <run_depend>geometry_msgs</run_depend>
-  <run_depend>libpcl-all-dev</run_depend>
+  <run_depend>libpcl-all</run_depend>
   <run_depend>libvtk</run_depend>
   <run_depend>message_runtime</run_depend>
   <run_depend>nodelet</run_depend>
-  <run_depend>opencv2</run_depend>
+  <run_depend>libopencv-dev</run_depend>
   <!--run_depend>pcl</run_depend-->
   <run_depend>pcl_ros</run_depend>
-  <run_depend>polypartition</run_depend>
+  <run_depend>libpolypartition</run_depend>
   <run_depend>roscpp</run_depend>
   <run_depend>rosbag</run_depend>
