--- polypartition/package.xml
@@ -1,18 +1,0 @@
-<package>
-  <name>polypartition</name>
-  <version>1.0.0</version>
-  <description>This package wraps the polypartition library to use it as a ros dependency</description>
-
-  <license>LGPL</license>
-
-  <url type="website">http://ros.org/wiki/polypartition</url>
-  <maintainer email="goab@ipa.fhg.de">Georg Arbeiter</maintainer>
-  <author>Georg Arbeiter</author>
-
-  <buildtool_depend>catkin</buildtool_depend>
-  <!--build_depend>mk</build_depend>
-  <build_depend>rosbuild</build_depend-->
-  <!--export>
-    <cpp lflags="-L${prefix}/lib -lphidget21" cflags="-I${prefix}/include"/>
-  </export-->
-</package>
